Fix size_t underflow in getRelnodeCoordinate path scan
The reverse scan for the second-to-last slash used a size_t index
with a 'start_off >= 0' condition, which is always true. On a
malformed path with fewer than two slashes the index would underflow
to SIZE_MAX and read out of bounds instead of hitting the error
branch. Use a signed index so the loop terminates correctly.

+++ src/util.cpp
@@ -40,7 +40,10 @@ relnodeCoord getRelnodeCoordinate(Oid spcNode, const std::string &fileName) {
 
   auto len = fileName.size();
 
-  size_t start_off = len - 1;
+  /*
+   * Scan backwards for the second-to-last '/'.
+   */
+  ssize_t start_off = (ssize_t)len - 1;
   int slash_cntr = 0;
   while (start_off >= 0) {
     if (fileName[start_off] == '/') {
